# cmd/api/docs/swagger.yaml
definitions:
...
  models.OrderFromCart:
    properties:
      address_id:
        type: integer
      payment_id:
        type: integer
    required:
    - address_id
    - payment_id
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact:
    name: API Support
  description: Show Times is an E-commerce platform to purchase Watch
  title: Go + Gin E-Commerce API Show Times
  version: 1.0.0
paths:
...
  /admin/orders:
    delete:
      consumes:
      - application/json
      description: Cancels an order based on the provided order ID from an admin perspective.
      parameters:
      - description: Order ID to cancel
        in: query
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: Order canceled successfully'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 'Internal server error: Error from orderID or couldn''t cancel
            the order'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Cancel order from admin
      tags:
      - Admin
    get:
      consumes:
      - application/json
      description: Retrieves all order details for admin with pagination based on
        the provided parameters.
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of items per page (default: 10)'
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: Retrieved all order details for admin successfully'
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 'Bad request: Page number or count not in correct format'
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 'Unauthorized: Invalid or missing authentication'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 'Internal server error: Could not retrieve order details for
            admin'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Retrieve all order details for admin
      tags:
      - Admin
    patch:
      consumes:
      - application/json
      description: Approves an order based on the provided order ID.
      parameters:
      - description: Order ID to approve
        in: query
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: Order approved successfully'
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 'Bad request: Error from orderID or couldn''t approve the order'
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 'Unauthorized: Invalid or missing authentication'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 'Internal server error: Failed to approve the order'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Approve order
      tags:
      - Order
...
  /user/orders:
    get:
      consumes:
      - application/json
      description: Retrieves order details for a user based on the provided pagination
        parameters.
      parameters:
      - description: User ID
        in: header
        name: id
        required: true
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of items per page (default: 10)'
        in: query
        name: count
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: Retrieved order details successfully'
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 'Bad request: Page number or count not in correct format'
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 'Unauthorized: Invalid or missing authentication'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 'Internal server error: Could not retrieve order details'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Retrieve order details
      tags:
      - Order
    patch:
      consumes:
      - application/json
      description: Initiates the return process for an order based on the provided
        order ID and user ID.
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Order ID to initiate return
        in: query
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: Order returned successfully'
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 'Bad request: Invalid order ID'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 'Internal server error: Couldn''t process the order return'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Initiate order return
      tags:
      - Order Management
    post:
      consumes:
      - application/json
      description: Places an order with items from the user's cart.
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Order details from cart
        in: body
        name: OrderFromCart
        required: true
        schema:
          $ref: '#/definitions/models.OrderFromCart'
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: Order placed successfully'
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 'Bad request: Error in getting user ID or invalid request'
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 'Unauthorized: Invalid or missing authentication'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 'Internal server error: Could not place the order'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Place order from cart
      tags:
      - Order Management
  /user/orders/checkout:
    post:
      consumes:
      - application/json
      description: Processes the checkout for the user's order.
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'Success: Checkout completed successfully'
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 'Bad request: Getting user ID failed'
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 'Unauthorized: Invalid or missing authentication'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 'Internal server error: Checkout failed'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Process checkout
      tags:
      - Order Management
...
securityDefinitions:
  BearerTokenAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"